Move `BitStream` to the appropriate module hierarchy
