+++ src/decoders/mod.rs
@@ -7,3 +7,4 @@ mod jpeg2000;
 mod png;
 mod run_length;
 mod simple;
+mod stream;
